backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
pytesseract==0.3.10
openpyxl==3.1.2